Unify and extract how to get hold of purs bin path
