+++ src/main/kotlin/org/purescript/ide/purs/StartupActivity.kt
@@ -10,16 +10,18 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.guessProjectDir
 import com.intellij.openapi.startup.StartupActivity
 import com.intellij.openapi.util.SystemInfo
+import java.io.File
 import java.nio.file.Path
 import java.util.concurrent.TimeUnit
 
 class StartupActivity : StartupActivity.Background {
     override fun runActivity(project: Project) {
-        val pursBin = getPursPath(project) ?: return
-        // node_modules/.bin/purs <- 3 parents
-        val rootDir = pursBin.parent.parent.parent.toFile()
+        // without a project dir we don't know where to start the server
+        val projectDir = project.guessProjectDir() ?: return
+        val rootDir = projectDir.toNioPath()
+        val pursBin = Purs().nodeModulesVersion(rootDir)
         val command = GeneralCommandLine(pursBin.toString(), "ide", "server")
-            .withWorkDirectory(rootDir)
+            .withWorkDirectory(rootDir.toFile())
         val task = object : Task.Backgroundable(
             project,
             "purs ide server for " + project.name,
@@ -32,24 +34,5 @@ class StartupActivity : StartupActivity.Background {
 
         }
         task.queue()
-    }
-
-    private fun getPursPath(project: Project): Path? {
-        val sequence = sequence<Path> {
-            var tmp = project.guessProjectDir()?.toNioPath()
-            while (tmp != null) {
-                yield(tmp)
-                tmp = tmp.parent
-            }
-        }
-        val nodeModules = sequence
-            .map { it.resolve("node_modules") }
-            .firstOrNull { it.toFile().exists() }
-            ?: return null
-        val binDir = nodeModules.resolve(".bin")
-        return when {
-            SystemInfo.isWindows -> binDir.resolve("purs.cmd")
-            else -> binDir.resolve("purs")
-        }
     }
 }
